Update form modal to support EspressoForm
