+++ assets/src/application/ui/forms/formModal/FormModalForm.tsx
@@ -1,22 +1,9 @@
-import React, { useEffect, useState, CSSProperties } from 'react';
+import React, { CSSProperties } from 'react';
 import { __ } from '@wordpress/i18n';
 import { ButtonProps } from 'antd/lib/button';
 import { Modal } from 'antd';
 
 import { FormModalFormProps } from './types';
-
-const formStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	padding: '1em 2em',
-};
-
-const dataStyle: CSSProperties = {
-	borderRadius: '5px',
-	boxSizing: 'border-box',
-	padding: '1em 2em',
-	color: '#a9ce47',
-	backgroundColor: '#26203d',
-};
 
 const modalStyle: CSSProperties = {
 	boxSizing: 'border-box',
@@ -32,28 +19,7 @@ const modalStyle: CSSProperties = {
 	transform: 'translate(-50%, -50%)',
 };
 
-const FormModalForm: React.FC<FormModalFormProps> = ({
-	FormComponent,
-	form,
-	values,
-	handleSubmit,
-	submitting,
-	pristine,
-	onClose,
-	...formProps
-}) => {
-	// boolean for communicating to form child components whether form needs to be reset
-	const [formReset, setFormReset] = useState(false);
-	// clear form data and set formReset back to false
-	// after form child components have had a chance to reset
-	// after receiving a positive formReset prop
-	useEffect(() => {
-		if (formReset) {
-			form.reset();
-			setFormReset(false);
-		}
-	});
-
+const FormModalForm: React.FC<FormModalFormProps> = ({ form, submitting, pristine, onClose, children }) => {
 	const submitButton: ButtonProps = {
 		disabled: submitting || pristine,
 		htmlType: 'submit',
@@ -70,7 +36,7 @@ const FormModalForm: React.FC<FormModalFormProps> = ({
 		htmlType: 'reset',
 		onClick: (click) => {
 			click.preventDefault();
-			setFormReset(true);
+			form.reset();
 		},
 	};
 
@@ -86,19 +52,7 @@ const FormModalForm: React.FC<FormModalFormProps> = ({
 			onCancel={onClose}
 			width={'80%'}
 		>
-			<form onSubmit={handleSubmit}>
-				<div style={formStyle}>
-					<FormComponent
-						form={form}
-						values={values}
-						submitting={submitting}
-						pristine={pristine}
-						formReset={formReset}
-						{...formProps}
-					/>
-				</div>
-				<pre style={dataStyle}>{JSON.stringify(values, null, 2)}</pre>
-			</form>
+			{children}
 		</Modal>
 	);
 };
