+++ assets/src/application/ui/forms/formModal/types.ts
@@ -1,9 +1,11 @@
 import React from 'react';
 import { FormProps, FormRenderProps } from 'react-final-form';
 import { FormApi } from 'final-form';
+import { EspressoFormProps } from '@application/ui/forms/espressoForm';
 
 export interface FormModalProps extends FormProps {
-	FormComponent: React.NamedExoticComponent;
+	FormComponent?: React.NamedExoticComponent;
+	formConfig?: EspressoFormProps;
 	onClose: (e?: any) => void;
 	isOpen: boolean;
 }
@@ -19,6 +21,6 @@ export interface FormComponentProps {
 }
 
 export interface FormModalFormProps extends FormRenderProps {
-	FormComponent: React.NamedExoticComponent<FormComponentProps>;
-	onClose: (e?: any) => void;
+	FormComponent?: React.NamedExoticComponent<FormComponentProps>;
+	onClose?: (e?: any) => void;
 }
